requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-redis==5.4.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2025.2
redis==6.2.0
requests==2.32.3
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
